refactor(logs): make the xterm.js screen the canonical LogsScreen
Rename ExperimentalLogsScreen to LogsScreen and the previous DOM-based
screen to LegacyLogsScreen, so the names match which renderer is the
default. Phase 3 only needs to delete the Legacy folder.
