+++ cmd/gomander/frontend/src/components/utility/EventListenersContainer.tsx
@@ -5,10 +5,11 @@ import { useTheme } from "@/contexts/theme.tsx";
 import { eventService } from "@/contracts/service.ts";
 import { Event, type EventData } from "@/contracts/types.ts";
 import { removeKeyFromLocalStorage } from "@/helpers/localStorage.ts";
+import { XTERM_THEMES } from "@/screens/LogsScreen/components/CommandTerminal.tsx";
 import {
 	formatLogTimestamp,
 	prependTimestamp,
-} from "@/screens/ExperimentalLogsScreen/helpers.ts";
+} from "@/screens/LogsScreen/helpers.ts";
 import { useCommandStore } from "@/store/commandStore.ts";
 import { terminalStore } from "@/store/terminalStore.ts";
 import { useUserConfigurationStore } from "@/store/userConfigurationStore.ts";
@@ -17,7 +18,6 @@ import { cleanCommandLogs } from "@/useCases/command/cleanCommandLogs.ts";
 import { detectMissingEnvironmentPathFailure } from "@/useCases/command/detectMissingEnvironmentPathFailure.tsx";
 import { recordCommandsErrors } from "@/useCases/command/recordCommandsErrors.ts";
 import { updateCommandStatus } from "@/useCases/command/updateCommandStatus.ts";
-import { XTERM_THEMES } from "../../screens/ExperimentalLogsScreen/components/CommandTerminal.tsx";
 
 export const EventListenersContainer = () => {
 	const addLogs = useCommandStore((state) => state.addLogs);
