+++ cmd/gomander/frontend/src/screens/LogsScreen/components/CommandTerminal.tsx
