+++ cmd/gomander/frontend/src/screens/LogsScreen/components/TerminalSearchBar.tsx
