+++ cmd/gomander/frontend/src/screens/LogsScreen/hooks/useTerminalSearch.ts
